requirements.txt
Django==2.0.4
django-haystack==2.8.1
django-tinymce==2.7.0
itsdangerous==0.24
jieba==0.39
Pillow==5.0.0
pkg-resources==0.0.0
PyMySQL==0.8.0
pytz==2018.4
Whoosh==2.7.4
